requirements.txt
# Core dependencies
transformers==4.46.0
peft==0.13.2
torch>=2.0.0
datasets>=2.14.0

# Geneformer (install from GitHub)
# Install with: pip install --no-deps git+https://huggingface.co/ctheodoris/Geneformer

# Data processing
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=12.0.0
anndata>=0.9.0
scanpy>=1.9.0
loompy>=3.0.0

# Machine learning
scikit-learn>=1.3.0
scipy>=1.11.0

# Optimization
optuna>=3.2.0
optuna-integration>=3.2.0
hyperopt>=0.2.7
ray>=2.5.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Training utilities
tensorboard>=2.13.0
tqdm>=4.65.0
statsmodels>=0.14.0
tdigest>=0.5.2

# Configuration
pyyaml>=6.0

# Optional: Weights & Biases (uncomment if needed)
# wandb>=0.15.0
